Minor Update + Discard 'Convert Screenshot' I'll handle this in C# because of GUI

+++ script/remove_empty_directory.py
@@ -2,8 +2,16 @@ import os
 import sys
 from argparse import ArgumentParser
 from pathlib import Path
+import stat
 
 from library.python_lib import Color, CustomFormatter  # noqa: E402
+
+
+exclusion = [
+    r"%AppData%\\Microsoft",
+    r"%LocalAppData%\\Microsoft",
+]
+exclusion = [Path(os.path.expandvars(path)).resolve() for path in exclusion]
 
 
 class RemoveEmptyDirectoryNamespace:
@@ -11,20 +19,36 @@ class RemoveEmptyDirectoryNamespace:
     yes: bool
 
 
-def remove_empty_directory(dir: Path, top_level: Path = None):
-    if top_level is None:
-        top_level = dir
+def remove_empty_directory(dir: Path, top_level: Path):
+    try:
+        if not dir.is_dir():
+            return
 
-    for subdir in dir.iterdir():
-        if subdir.is_dir():
+        is_junction = stat.S_ISLNK(os.stat(dir, follow_symlinks=False).st_mode)
+        is_symlink = dir.is_symlink()
+
+        if is_symlink or is_junction:
+            print(f"{Color.BLUE}SYMLNK{Color.RESET} {dir}")
+            return
+
+        for subdir in dir.iterdir():
+            if subdir.is_dir():
+                if any(subdir.resolve().is_relative_to(exc) for exc in exclusion):
+                    print(f"{Color.BLUE}SKIP  {Color.RESET} {subdir}")
+                    continue
             remove_empty_directory(subdir, top_level)
 
-    if dir != top_level:
-        try:
-            dir.rmdir()
-            print(f"{Color.YELLOW}REMOVE{Color.RESET} {dir.parent}{os.sep}{Color.YELLOW}{dir.name}{Color.RESET}")
-        except OSError:
-            pass
+        if dir != top_level:
+            try:
+                dir.rmdir()
+                print(f"{Color.YELLOW}REMOVE{Color.RESET} {dir.parent}{os.sep}{Color.YELLOW}{dir.name}{Color.RESET}")
+            except OSError:
+                pass
+
+    except PermissionError:
+        print(f"{Color.RED}PERMERR{Color.RESET} {dir}")
+    except NotADirectoryError:
+        pass
 
 
 ##########
@@ -50,4 +74,4 @@ if not args.yes:
         print(f"{Color.RED}ERROR{Color.RESET} User canceled the opration")
         sys.exit(1)
 else:
-    remove_empty_directory(args.target)
+    remove_empty_directory(args.target, args.target)
